Sync upstream (#10)
* fix: Misfunction of wechat_qr in Social Integration (#2831)

Solve the bug issue (#2830)

* Support for openpanel.dev for analytics (#2822)

* Blog Page Navigation (#2841) (#2846)

I tested on my Github page and it worked correctly. Then, copied here.

Issue: Blog page navigation extends layout of the page in small screens.
Fixes: #2841

* Bluesky URL prefix should not be appended (#2856)

This PR modifies

- `_config.yml` to rename `bluesky_url` into `bluesky_id`
- `_includes/social.liquid` to handle the renaming in the display of
social media icons

* Page permalink can be an outside url (#2863)

Fixes #2858

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Fixed bug in header

* Fix prettier complaints

* Fix to display references in order of citation (#2855)

"cite_in_order" here contradicts the global group_by year setting in
_config.yml

* Add .lycheeignore file with linkedin to stop error on workflow Linkcheck (#2868)

Add a .lycheeignore file with LinkedIn to stop the error in the workflow
Link Checker due to response 999, and update to the new version
lycheeverse/lychee-action@v2.1.0.

* Support for Resume/CV Links (#2826)

Solves #2825

* Add features to distill template (#2840)

While using the distill template, I noticed it lacked many features
available in the regular post template, such as typograms, echarts, and
other interactive elements. To address this, I **included the necessary
scripts and dependencies for these features in** _distill.liquid_,
referencing them from misc.liquid. Now, the distill template offers the
same interactive functionality as the post template.

* Fix apostrophe in news title for search (#2879)

Fix #2876

Weirdly enough, we already escape the title a few lines above the
changed one, but had to escape it again here.

Signed-off-by: George Araújo <george.gcac@gmail.com>

* New image components (image galleries) (#2877)

Added support for multiple image galleries, and created a sample post
showing how to enable and use them.

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Update Font Awesome and Tabler Icons (#2884)

Solves #2881.

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Moved socials config to `_data/socials.yml` (#2885)

Fixes #2882, added support for
[scholar-icons](https://github.com/louisfacun/scholar-icons) and
[LeetCode](https://leetcode.com/).

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Renamed `repo-img-` to `only-` to allow setting images for only one theme (#2886)

Closes #930.

We already had classes for this kind of feature, but it was used
specifically for repository information. Just renamed the classes so it
is clear how they can be used.

To use them with markdown, just do it like this:

```markdown
![Only light](assets/img/prof_pic_color.png){: .only-light}
![Only dark](assets/img/prof_pic.jpg){: .only-dark}
```

and with html:

```html
<img
  class="only-light"
  src="assets/img/prof_pic_color.png"
>
<img
  class="only-dark"
  src="assets/img/prof_pic.jpg"
>
```

Signed-off-by: George Araújo <george.gcac@gmail.com>

* Update docker-compose.yml image version

* Updated dependencies (#2887)

Signed-off-by: George Araújo <george.gcac@gmail.com>

---------

Signed-off-by: George Araújo <george.gcac@gmail.com>
Co-authored-by: Hanyan Yin <yinhanyan@outlook.com>
Co-authored-by: Abderrahmane Ait gueni ssaid <45363061+aaitguenissaid@users.noreply.github.com>
Co-authored-by: imsi32 <81227251+imsi32@users.noreply.github.com>
Co-authored-by: Guillaume Dalle <22795598+gdalle@users.noreply.github.com>
Co-authored-by: George <31376482+george-gca@users.noreply.github.com>
Co-authored-by: Juan Carlos Niebles <niebles@gmail.com>
Co-authored-by: Isaac Fernandes <101341132+Eyezuhk@users.noreply.github.com>
Co-authored-by: Pranjal Aggarwal <54495980+pranjalagg@users.noreply.github.com>
Co-authored-by: dai <79570336+ShuhongDai@users.noreply.github.com>